product-service/requirements.txt
flask
pymysql
sqlalchemy
google-cloud-sqlcommenter
protobuf
